browsedb.pl v. 0.86
browsedb.pl v. 0.86
The display and handling of errors, especially in the
dbitable module was improved.
Inserting and deleting of a line should now work.
A simply copy and paste functionality was added.
The table-windows have no longer a "close" button, you
can just close them in the standard way by double clicking on
the upper left corner.