The table-selection dialog was vastly improved. It
The table-selection dialog was vastly improved. It
now has a drop-down list and kind of command completion.
However: there is still a bug in finding valid table and
view-names. And check_existence in dbdrv_oci.pm
needs to be changed. For testing, in line 10 in dbdrv.pm just
activate the return statement, then everything should work.