there is a new option "no_auto_pk"
there is a new option "no_auto_pk"
with this, a "0" in the primary key column is
not used to create a new primary key. Instead, the
user has to specify all primary keys by himself.
This is especially useful, if the table has an
entry where the primary key is 0.