colum.maps have been introduced. By this any SQL statement can
colum.maps have been introduced. By this any SQL statement can
be used to map values in the table to other values, just by taking
the first two columns of the result of the SQL statement.
This is escpecially useful for foreign keys.
The usage of the colum-map can easily switched on and off
by the column popup-menu.