the load function (load from a database) now does not always
the load function (load from a database) now does not always
call dbdrv::error(), especially when the SQL statement
was faulty. The interal  hash-element "_fetch_cmd" is now
reset to it's old value, when the new SQL command had an error.
When the SQL command fails, data already present in the dbitable-object
is preserved.