this script shows the difference between a local copy of
this script shows the difference between a local copy of
an capfast file and the version in the repository. It uses
Sch2db.pl to convert the files to the db-format and then
dbdiff to display the differences. By this, only the differences
that really matter are shown. Modifications like graphical re-arranging
of the records, are not shown.