parse_subst has been completely rewritten. It
parse_subst has been completely rewritten. It
now uses the same parsing techniques as parse_db.pm.

The string to parse can now be given as a scalar or
a reference to a scalar.

In case of an error, the line-number and column number
of the position in the parsed text is also shown.