some core parts of expander were re-written.
some core parts of expander were re-written.
The module now uses constants instead of strings
at some places in order to make the parsing process
faster.

As a result, the scanning of round-bracket variables or
the refusal of non-bracketed variables is now
more consistant. Dollar-expressions that are
not recognized are now left unchanged and there is
an option that causes the parser not to exit but only
to warn when there is an unknown variable found.

The searching of paths in order to find files of the
$include statement is now done much faster, since the
program stays in the last path, where a file was
successfully found. This may, but should usually not, cause
problems in some applications of expander.pm

Two options are new:
allow_not_defined_vars and
forbit_nobrackets