This module scans one or more than one makefile and returns
This module scans one or more than one makefile and returns
a hash reference containing all variables that are set within
the makefile together with all environment variables.

In order to speed up the scanning of the makefile, the program
also provides a function that creates and manages a cache-file
that can be read more quickly than scanning the makefile again
and again.