pcomp.pl was added,
pcomp.pl was added,
this tool recursively compares directories.
It shows files and directories that do not
exist in both directories, files with different
dates, different sizes or different contents.
The new terse mode (-t) print a each file in a single
line with a set of flags prepended similar to the
output of "cvs -n -q update".