rec_link_fields() didn't work correctly when
rec_link_fields() didn't work correctly when
the record had it's DTYP field not defined but was referring
to another record via it's OUT field. analyse_db suppresses
references when DTYP is unequal to "soft channel". Now we
assume that a DTYP that is not defined is also "soft channel".