rsync-linklog-stat.py was added.
This script analyses the log-links file that
rsync-dist.pl produces. It uses the new --python-log
feature of rsync-dist.

This tool can also be used to find out which versions
in a distribution directory are no longer used or were
just test versions that were used for a very short time.
It can list these versions, which then may be manually
be completely removed from the distribution directory.