a small improvement in the program's output
Before, with "-n" format, the star "*" marking
the active version could be printed more than once
if a version is more than once in the list. Since this
may be confusing this has been changed. The star is now
printed only on the last occasion the active version is
in the list.