bugfix in rsync-dist.pl, --checksum didn't work as expected.
In the old version of rsync-dist.pl, when --checksum was
used, this option was passed to rsync together with rsync's
update option.
The effect was that only files that were newer at the source
and whose checksum was different from that of the destination
site were transferred.

The desired way of operation was, however, that ALL files
different at the source should be transferred to the
server, regardless of their file date.

In this new version of rsync-dist.pl, the program should now behave as
expected when --checksum is used.