rsync-dist-info is an improved version of rsync-linklog-stat
rsync-dist-info is the successor of rsync-linklog-stat. The
program uses mostly the same algorithms as the old script,
but much of it's internal functions have been moved to
external python modules. An importtant new function is
the "--boot-time" command which is used to compare
version activation times with the boot time.