support for scripts that need python 2.5 was added
Some changes in Makefile:

python scripts that do not run with python 2.4 now call
"python2.5" in their first script line if this binary
was found during execution of make. Otherwise
"python" is called in the first script line. This is needed
since on host aragon, "python" is python 2.4 and "python2.5"
is python 2.5.