bugfix: the creation of the recovery file sometimes failed.
The mechanism to exclude the recovery data from beeing put into
the tar file of files unknown to the repository failed under
certain cicumstances (when the recovery directory is below a sub-directory
of the working copy directory, but in a sub-directory that is itself
not under version control). "hg status -X [exclude-file]" fails if
[exclude-file] is in a sub-directory not under version-control.

Not hg-recover.py first gets all files unknown to the version control
system and removes the recovery directory later.