better error handling when documentation is generated
The generation of python documentation requires that the python scripts and
modules can be executed. Since some of them use python modules that are part of
bii_scripts, the location of them has to be prepended to the python module
search path. This is now done for all calls to the python interpreter and for
calls to pydoc. 

However, when modules that are not part of bii_scripts are prerequisites,
running the scripts or generating documentation with pydoc fails. In order to
handle this, all error messages are collected in an error log file "ERRLOG.TXT"
that is placed in the "out" directory. The generation of other documentation
files will continue, but the user will be warned and the ERRLOG.TXT file is
printed to the screen if it is not empty.