hgdiff was replaced by hgdiff.py, a much improved version of hgdiff
Since hgdiff was a simple shell script, it had many limitations
and was a bit complicated to use. hgdiff.py is a re-write of this
script in python with much improved functionality. It is, however,
still a simple browser for file differences that uses tkdiff and
has no ambitions to become a GUI frontend for mercurial.

Hope it is useful...