bugfix in dbdrv_oci.pm: "schema.tablename" didn't always work.
In function real_name() a database object of the form "schema.tablename" was
treated as inaccessible if it was not in the object dictionary for the given
user. However, it is possible that an object not in the object dictionary is
accessible.

Now, if the given name is not found in the object-dictionary for the user and
the name has the form "schema.tablename", it is assumed that this IS accessible
for the user anyway.