bugfix in expander.pm: for-loops where condition is initially false.
In the previous version of expander.pm, for-loops were always run once, even if
the condition was false from the start. Since this behavior is different from
most programming languages, it has been changed. Now, when the for-condition is
initially false, it is treated similar to an if-block with a false condition.
Everything between "for" and "endfor" is skipped in this case. This also
applies to "for_noblock".