The distribution of documentation files was adapted to the new www-csr server.
If this makes problems during "make install", simply set
"USE_RSYNC=no" in the Makefile.

"make install" now distributes the documentation and the sources of bii_scripts
to www-csr.bessy.de with the rsync tool.

The file CONTENTS.txt was changed for the slightly different file structure on
the web server.