Bugfix in hg-recover, it now uses hashkeys instead of revision numbers.
The revision numbers may change, if the central repository has changed from the
time the recover data was created to the time the recover data is used to
restore a repository. This causes hg-recover sometimes to end with an
exception.

This new version now uses the unique hash-keys to identify patches, which
should avoid such problems. It is, however, backwards compatible with recover
data created with the old version of hg-recover.