Python file documentation now works even if some needed modules are missing.
The automatic documentation generation for python scripts and modules is only
possible when these files can be read and executed by the python interpreter
without errors.

But when only one of the imported python modules is missing, python raises an
exception and stops interpretation of the file so the documentation could not
be created. With this patch, some of the python scripts and modules have been
modified in a way that the missing of some non standard python modules is
tolerated when only documentation is to be generated. In this case, a warning
is printed to stderr but the interpretation of the file continues.

The out/ERRLOG.TXT file that is created during make now may contain warnings
with respect to missing python modules. All these lines start with "WARNING:".
When make checks if there were any error messages, it ignores lines starting
with "WARNING:" in the errlog file.

In all scripts modified by this patch, tabulators were removed in the source.