BDNS.pm was improved, BDNS.py was added.
New functions were added to BDNS:
    parse_named : returns the parse results as a hash(perl) or dict(python)
    sortNamesBy : allows the sort order to be specified as a parameter instead of
                  a global variable.
    mkOrder     : create a sort order array (for sortNamesBy) from a given
                  string or list of strings/numbers.
    cmpNamesBy  : compare two devicenames for a given sort order array.

Additionally, BDNS.py was added. It has the same functionality as BDNS.pm but
is written in python. The original BDNS.pm was changed for the different syntax
in Python with some minor modifications.

Testcode was added in the new directory "tests":

Compare BDNS.pm and BDNS.py like this:

    cd tests/BDNS
    cat PVLIST | perl test_bdns.pl > PERL
    cat PVLIST | python test_bdns.pm > PYTHON
    diff PERL PYTHON

Both files should be equal!