hg2darcs.py was added, this converts mercurial to darcs revisions.
With this script it is possible to get a working copy from a darcs
repository, commit all changes in a local mercurial repository and push back
the changes to darcs.

In order to do this you have to create a mercurial repository in your working
copy with "hg init". Then you can do your local software development in
mercurial. When it is time to publish the changes back to the central darcs
repository, hg2darcs.py can convert all or some of the changes recorded in
mercurial back to the local darcs repository. These can then be sent to the
central darcs repository with "darcs push".

hg2darcs.py keeps the changes, the log message and the author information
intact, only the record date changes to the time hg2darcs.py was started.