The embedded help was corrected with respect to RSYNC_DIST_DEREFERENCE_LINKS.
The option that is used to deference symbolic links when files are
distributed is named --dereference-links. Although the appropriate constant
in the configuration file was named RSYNC_DIST_DEREFERENCE_LINKS, it was
referred as RSYNC_DIST_REMOVE_LINKS in the documentation and the generated
comment lines in the configuration file. This has now been fixed.