Bugfix in expander.pm: columns() returned garbage when used with $include.
When $include was used in an expander file, the column-numbers were
calculated wrong after the place in the file where the "$include" statement
was placed. This was due to the fact that the column calculation mechanism
used global variables. This collides with the implementation of $include,
which calls parse_file() recursively.

Now the global variables have changed to local variables in the function
parse_scalar_i(). The global functions column() and lineno() use global
references to the column and line numbers.

Note that with this implementation, line and column numbers are calculated
for each local file. So when a file is included with "$include", the line
numbers start to count again with 0.