bugfix in rsync-dist.pl: the version file was sometimes not distributed.
The symptom:
In the BII-Controls and HoByCat project, distributing a version with:
  make rsync-dist.IOC FILES=<filename>
sometimes creates a distribution directory where the file "iocBoot/version"
is still of the old version, meaning it's newer version was not distributed.
With the new startup scripts that use the content of the version file to
change their directory after boot, they end up in the wrong directory with an
old software version.

The explanation:
rsync-dist.pl is called with the "--partial" option in this case. The
internal function "dist" creates a temporary file with all files that are
given to the "--localpath" option. In order to support file glob patterns, it
searches for the files or directories. The file "iocBoot/version" is created
with the "--version-file" option, but AFTER the list of files to distribute
is computed. If the file "iocBoot/version" doesn't exist because the
development tree was just created by a "darcs get", the file-search routine
doesn't find the version file. In this case the file is missing in the list
of files to transfer, so it is not distributed. In the next run, when the
version file is found since it was created in the last call of rsync-dist.pl,
it IS FOUND and distributed as it should. So the effect only takes place when
a "--partial" is used on a development tree where rsync-dist.pl was never
called before.

The fix:
in function dist(), the version file is now created BEFORE the list of files
to transfer is created.