tableutil.py was added, a tool to do calculations with tables of numbers.
This script can be used to manipulate and pretty-print tables of numbers. Such
a table consists a line that defines the names of the columns and more lines
that contain numbers separated by spaces. It can add new columns by applying
calculations on the existing columns as well as execute command scripts (using
python and numpy_table.py) to do more complex operations on tables.

Two modules, numpy_util.py and numpy_table.py were added that are needed by the
script. This script requires the python numpy module to be installed.