Python3 support was added to the Makefile.
When CREATE_PYTHON3 is set to "yes", the Makefile now creates python3
versions of all python modules and some of the python scripts. This is done
by converting the files with the python3-2to3 utility which is part of the
python3 distribution. The modules are copied to the directory
out/lib/python3. Converted scripts have a "3" inserted before the "." in
their filename and are stored among the other scripts in out/scripts.