Python modules and scripts were changed for python3.
Python3 is (slightly) incompatible with python2. There is an converter tool,
python3-2to3, that can convert a python2 to a python3 file, but some things
have still to be changed manually.

Currently, all python development in this project is done for python2 but we
try to make all python files automatically convertable to python3 by usage of
the python3-2to3 script. When CREATE_PYTHON3 in the Makefile is set to "yes",
it creates python3 versions of all python modules and some of the python
scripts.

Due to these considerations, some changes were necessary in some of the
python files which are combined in this patch.

It was ensured, that all the doctest testcode in the files can also run in
the python3 versions without errors.