Better error handling in numpy_table.py and numpy_util.py.
Numpy knows several kinds of numerical exceptions. However, the default
behavior of Numpy is to just print a warning and continue. Since this warning
doesn't show the exact line number where the error occured, this makes finding
the error very tedious. 

The function "seterr" of Numpy can be used to change this behavior in a way
that an error raises an ordinary python exception. 

numpy_util.py now sets this mode, all Numpy errors are python exceptions. If
the user wants to change this, he can call numpy_seterr which directly calls
seterr from the Numpy module. This function is also available in the module
numpy_table.py.