Bugfix in hg-recover.py: qparent must be taken into account.
When a repository with mq patches is restored, the program needs to know what
the "qparent" version was when the recovery data was generated. This is now
also added to the metadata file. When the repository is generated, the "hg
clone" command clones the central repository up just to the qparent version.
On this version the mq patches are applied.