hg-recover.py can now restore some repositories with which it failed before.
In order to recover the state of the mercurial patch queue, the script needs
to know the "qparent" revision. In the older versions of recover files this
information was not stored, causing hg-recover.py to fail in restoring these
repositores. Now the script can restore the "qparent" version if it is
missing in the restore file, although the method is a but complicated...