hg-sig was added, this helps comparing mercurial repositories.
This script generates a text file for a mercurial repository which contains
all the log messages, node IDs and possibly checksums but not dates or
authors. This can be used to compare mercurial repositories especially when
mq patches are used.

The script can also directly call a graphical diff viewer program, e.g.
tkdiff.