hg-recover.py now also handles phases.
When hg-recover.py tries to restore the mq patch queue, it has to convert
mercurial patches to mq patches. This is by default forbidden by the the
"phases" feature or mercurial. Now the script uses "phase --force" to make the
patches editable. The program also handles the case when it is called with a
version of mercurial that doesn't know phases.