In boottime.py, the U125/1 undulator was added.
Currently the module contains a datastructure that maps "idcp<nn>" to
devicenames for all undulators here. This is a but unelegant since this
information would be available in StructuredData. However, it is not clear if
it would be a good idea to make this script dependent on finding the undulator
StructuredData parameter file. So for now, we just added the U125/1 to the
internal list.