Bugfix in boottime.py, U49-2 boottime couldn't be found.
The U49-2 undulator is the first undulator with the new EPICS statistic
records. Like all the other IOCs in the BESSY controlsystem this IOC now has a
"bootTime" record. bootTime.py was changed that is first tries to get the value
of record "bootTime" for an insertion device, only when this fails it tries to
read the record "uptime".