parse_db::parse_file can now parse more than one file.
Function parse_file in parse_db.pm has been modified in a way that is still
compatible with it's old function. If the first parameter is a string, it
simply parses that one db file. If the first parameter is a reference to a list
of strings, it parses all the given db files and returns the results in a
single structure.