dbfilter.pl can now combine several db files.
Before this patch, dbfilter when given more than one filename used to parse and
analyse each file separately. This behavior was often not what was expected.
Now, if several files are specified on the command line, they are all parsed as
if their contents were given in a single file.