dbfilter.pl now takes link fields from analyse_db.pm.
The list of link fields in a record is now taken from analyse_db.pm. The
variables @dtyp_link_fields and @link_fields in this module are now exported.

In the future, analyse_db should take the list of link fields from a dbd file
instead of having it hard-coded there.