dbfilter.pl -R now displays relation details.
Option -R or --record-references now displays the not only related records but
the fields that produce the relation along with the target fields and flags
like "CPP" or "CA". This is taken from the link labels of the "--dot" report.

Note that the alternative format for record references that was enabled with
oprion "-a" has been removed with this patch. The improved possibilities to
filter results with regular expressions make this one-liner format obsolete.