Bugfix in hg2darcs.py.
darcs stopped when we tried to record moves of several files. The script moved
a file to it's old location in order to let darcs move it to the new location.
Somehow this was a problem for darcs and it showed that the reverse move of the
file is actually not (no longer) needed in darcs. 

With the reverse file move was removed now everything works as expected.