The setting of sys.path in rsync-dist-info.py was improved.
The library search path in modified in this script due to a misconfiguration on
host elbe. Before this patch, a new search path was appended to the front of
the python search path. This however, lead to unexpected behaviour when
PYTHONPATH was changed for program tests. Now the part of the library search
path that causes the problems is removed.