Bugfix in rsync_dist_lib.py, the date was not always parsed correctly.
Since dates printed by "ls -l" not always contain a year we need a hint what
this year could be. Fortunately, each entry in file LOG-LINKS contains a field
"DATE" which is an ISO date. The year from this field is taken if the year in
the entry of "ls -l" is missing.