In maillike.py, wrong input data is now recorgnized.
The program assumes that the first non empty line that is not equal to "%%"
has the form:

  NAME: VALUE

Where NAME is a string consisting of letters, digits or the "_" character.
Also, NAME must not entirely consist of "_" characters and must be non-empty.

If this is not the case, the programs raises a ValueError exception and shows
the text it tried to parse.