rsync-dist.pl now allows aribtrary prefixes for link-paths.
Option RSYNC_DIST_PREFIX_DISTDIR was extended in it's function.

It may be "" or "1" when the user wants the symbolic links to refer to the
absolute patch of the distribution directory.

It may also be an abitrary path which is then used to create the path the
symbolic links refer to. For example, if RSYNC_DIST_PREFIX_DISTDIR is set to
"../dist" the symbolic links created by rsync-dist.pl refer to this relative
path instead of an absolute path. This is needed for modern tftp servers that
do not allow symbolic links with absolute paths.