The file SETENV-RUN-LOCAL.sh now checks the current working directory first.
This script modifies the environment variables PATH, PERL5LIB and PYTHONPATH so
that scripts in the bii_scripts source tree can be run directly. For this it
must know the locatation of the bii_scripts source tree which it takes from the
current working directory. If the bii_scripts source directory is not found
there, the script aborts.

The script must be *sourced* not *executed* like here:

  . ./SETENV-RUN-LOCAL.sh