idcp-restore: The script no longer aborts on hg-recover.py errors.
This is in order to have the id_db database restored even if restoring idcp
failed. As from the middle of 2019 on, a change in mercurial caused
hg-recover.py to not include the patch queue in the recover data. So recover
files from mid 2019 to mid 2020 cannot be extracted without errors. Fortunately
in most cases it is possible to restore versions with the help of the mercurial
signature file (see "hg-sig.py") manually.