iddb was partially rewritten and now supports reading a StructuredData file.
With option "-f FILENAME" the script reads a StructuredData file instead of
contacting the XMLRPC server. 

Only when this option is used, the program loads the StructuredData python
libraries so for this command, StructuredData must be installed.

If option "-f" is not used, this program can is still doesn't require
StructuredData to be installed on your local system. In this case it only uses
python standard libraries.