repo-mirror.sh: Allow to specify the .hgrc file.
Since the hggit extension, which is activated in the hgrc file, breaks pulling
repositories by http, hggit is not activated in my personal hgrc file.

For now, the user of repo-mirror.sh can specify a hgrc file at some othet
location which has hggit enabled just for the purpose of mercurial --> git
conversion.