added script to "fix" no-alarm-colors in edl files
Originally, we defined a no-alarm-color for edm to use, when a color is set 
to be "Alarm Sensitive". This changed the defaut behavior of edm and produced 
conflicts with edl files coming from external sources.

Ths script is to fix these issues by setting any color marked as 
alarm sensitive to the default NO_ALARM green.